Move build from build/dist to lib
- This is a more traditional location and it fixes problems with the
location of package.json
