+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules
 /.nyc_output/
 /coverage/
-build/dist/
+lib
