+++ package.json
@@ -12,7 +12,7 @@
     "format": "prettier --write '**/*.js'",
     "test": "npm run coverage",
     "test-unit": "jest --env node",
-    "compile": "rm -rf build/dist/* && tsc",
+    "compile": "rm -rf lib/* && tsc",
     "browserify": "npm run compile; grunt browserify"
   },
   "dependencies": {
