+++ src/deprecate.ts
