+++ src/has.ts
