+++ src/internal_config.json
@@ -1,0 +1,4 @@
+{
+    "INITIAL_RETRY_DELAY_IF_RATE_LIMITED": 5000,
+    "MAX_RETRY_DELAY_IF_RATE_LIMITED": 600000
+}
