+++ src/query.ts
